+++ kommons-uri/src/commonMain/kotlin/com/bkahlert/kommons/uri/StringValues.kt
@@ -2,6 +2,7 @@ package com.bkahlert.kommons.uri
 
 import io.ktor.util.StringValues
 import io.ktor.util.StringValuesBuilder
+import kotlin.reflect.KProperty
 import io.ktor.http.formUrlEncode as ktorFormUrlEncode
 
 /**
@@ -31,3 +32,12 @@ public fun StringValues.formUrlEncode(keepEmptyValues: Boolean = true): String =
             listOf(key to null)
         }
     }.ktorFormUrlEncode()
+
+/**
+ * Returns all values of the parameter with the name of the delegates [property]:
+ * - each value corresponds to one parameter occurrence (for example `name=value1&name=value2`)
+ * - an empty list represents a parameter with no value (for example `#name`)
+ * - `null` represents no parameter with name at all
+ */
+public operator fun StringValues.getValue(thisRef: Any?, property: KProperty<*>): List<String>? =
+    getAll(property.name)